stop serializing motd and starting message packets on each sent packet
